requirements-api.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
requests>=2.31.0
python-dotenv>=1.0.1
python-multipart>=0.0.9